+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next
 
 - Compatibility with PHP 8.0
+- Rate limit: always log if the configured limit was reached, even if it's not active.
 
 ## 1.15.2